+++ src/changelog.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Arma 3 Balaclavas
   - Arma 3 Bandanas
   - MA: TGE Balaclavas
+- Luther Custom JFO helmet
+- 34th IHADSS Helmet
 
 ## 0.22.0
 ### Added